+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.45.0 (2025-mm-dd)
+## 1.45.0 (2025-01-27)
 
 - **ENH: Use `odc.geo.xr.crop` instead of `rio.clip` to make `rasters.crop` dask-compatible** ([#27](https://github.com/sertit/sertit-utils/issues/27))
 - ENH: Allow to pass the version number as a string to `misc.compare_versions`